+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Please add changes at the top. When releasing a version add a new header for tha
 
 ### master
 
+ - Two new attribute operations implemented: _Fill edge attribute with constant default value_
+   and _Merge two edge attributes_. These do the same to edges that
+   _Fill with constant default value_ and _Merge two attributes_ could do to vertices. Note that
+   the latter two operations have been renamed to
+   _Fill vertex attribute with constant default value_ and _Merge two vertex attributes_,
+   respectively.
  - Lin's Centrality algorithm is added.
 
 ### 1.4.1