+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Please add changes at the top. When releasing a version add a new header for tha
 
 ### master
 
+ - Improvements to the import graph workflow: somewhat less computation stages plus the
+   algorithm is not sensitive anymore to vertices with millions of edges.
  - Two new attribute operations implemented: _Fill edge attribute with constant default value_
    and _Merge two edge attributes_. These do the same to edges what
    _Fill with constant default value_ and _Merge two attributes_ could do to vertices. Note that